# service/apigateway/docs/swagger.yaml
definitions:
...
  response.ApiResponseMerchantDocument:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/response.MerchantDocumentResponse'
        description: Single document response
      message:
        description: Descriptive message
        type: string
      status:
        description: Response status
        type: string
    type: object
...
  response.ApiResponsePaginationMerchantDocument:
    properties:
      data:
        description: Array of document data
        items:
          $ref: '#/definitions/response.MerchantDocumentResponse'
        type: array
      message:
        description: Descriptive message
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/response.PaginationMeta'
        description: Pagination metadata
      status:
        description: Response status
        type: string
    type: object
  response.ApiResponsePaginationMerchantDocumentDeleteAt:
    properties:
      data:
        description: Array of documents (with deletion info)
        items:
          $ref: '#/definitions/response.MerchantDocumentResponseDeleteAt'
        type: array
      message:
        description: Descriptive message
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/response.PaginationMeta'
        description: Pagination metadata
      status:
        description: Response status
        type: string
    type: object
...
  response.ApiResponseTransactions:
    properties:
      data:
        description: Array of transaction data
        items:
          $ref: '#/definitions/response.TransactionResponse'
        type: array
      message:
        description: Response message
        type: string
      status:
        description: Response status
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        description: HTTP status code or application error code
        type: integer
      message:
        description: Human-readable error message
        type: string
      status:
        description: Error status ("error", "fail", etc.)
        type: string
    type: object
  response.MerchantDocumentResponse:
    properties:
      created_at:
        description: Timestamp when document was uploaded (RFC3339 format)
        type: string
      document_type:
        description: Type of document (e.g., "license", "identity_proof")
        type: string
      document_url:
        description: URL or storage path where the document is saved
        type: string
      id:
        description: Unique document identifier
        type: integer
      merchant_id:
        description: ID of the merchant this document belongs to
        type: integer
      note:
        description: Administrative notes about the document
        type: string
      status:
        description: Current status ("pending", "approved", "rejected")
        type: string
      updated_at:
        description: Timestamp when document was last updated
        type: string
    type: object
  response.MerchantDocumentResponseDeleteAt:
    properties:
      created_at:
        description: Original creation timestamp
        type: string
      deleted_at:
        description: Deletion timestamp (nil if not deleted)
        type: string
      document_type:
        description: Type of document
        type: string
      document_url:
        description: Document storage path/URL
        type: string
      id:
        description: Unique document identifier
        type: integer
      merchant_id:
        description: ID of the merchant
        type: integer
      note:
        description: Final administrative notes
        type: string
      status:
        description: Last known status before deletion
        type: string
      updated_at:
        description: Last update timestamp
        type: string
    type: object
...
  response.PaginationMeta:
    properties:
      current_page:
        description: Current page number (1-based index)
        type: integer
      page_size:
        description: Number of items per page
        type: integer
      total_pages:
        description: Total number of pages available
        type: integer
      total_records:
        description: Total number of items across all pages
        type: integer
    type: object
...
  response.TransactionResponse:
    properties:
      amount:
        description: Transaction amount
        type: integer
      card_number:
        description: Card number used for the transaction
        type: string
      created_at:
        description: When the record was created
        type: string
      id:
        description: Unique transaction identifier
        type: integer
      merchant_id:
        description: ID of the merchant where transaction occurred
        type: integer
      payment_method:
        description: Payment method used (e.g., "credit", "debit")
        type: string
      transaction_no:
        description: Unique transaction reference number
        type: string
      transaction_time:
        description: Timestamp of the transaction
        type: string
      updated_at:
        description: When the record was last updated
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/merchant-document-query:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all merchant documents
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of merchant documents
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchantDocument'
        "500":
          description: Failed to retrieve merchant document data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all merchant documents
      tags:
      - Merchant Document Query
  /api/merchant-document-query/{id}:
    get:
      consumes:
      - application/json
      description: Get a merchant document by its ID
      parameters:
      - description: Document ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Document details
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDocument'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to get document
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Get merchant document by ID
      tags:
      - Merchant Document Query
  /api/merchant-document-query/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all active merchant documents
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active merchant documents
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchantDocument'
        "500":
          description: Failed to retrieve active merchant documents
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all active merchant documents
      tags:
      - Merchant Document Query
  /api/merchant-document-query/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all trashed merchant documents
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed merchant documents
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchantDocumentDeleteAt'
        "500":
          description: Failed to retrieve trashed merchant documents
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all trashed merchant documents
      tags:
      - Merchant Document Query
...
  /api/transactions-query/merchant/{merchant_id}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of transactions using the merchant ID
      parameters:
      - description: Merchant ID
        in: query
        name: merchant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction data
          schema:
            $ref: '#/definitions/response.ApiResponseTransactions'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve transaction data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find transactions by merchant ID
      tags:
      - Transaction
...
swagger: "2.0"